cpp: test: Split TIFF sample discovery from tiff tests
This permits the samples to be reused in other tests.
