+++ cpp/test/ome-bioformats/CMakeLists.txt
@@ -113,7 +113,7 @@ if(BUILD_TESTS)
 
   bf_add_test(ome-bioformats/planeregion planeregion)
 
-  add_executable(tiff tiff.cpp)
+  add_executable(tiff tiff.cpp tiffsamples.cpp)
   target_link_libraries(tiff ome-bioformats)
   target_link_libraries(tiff ome-test ${PNG_LIBRARIES})
   add_dependencies(tiff gentestimages)
